+++ src/models/litho/lens_torch.py
@@ -62,7 +62,7 @@ class Lens:
         H = torch.where(H > 1.0, 0.0, 1.0)
         R[R > 1.0] = 0.0
 
-        W = torch.zeros((2 * self.gnum + 1, 2 * self.fnum + 1), dtype=torch.complex128)
+        W = torch.zeros((2 * self.gnum + 1, 2 * self.fnum + 1), dtype=torch.complex64)
 
         for ii in range(len(self.Zn)):
             W = W + zerniken(self.Zn[ii], R, TH) * self.Cn[ii]
